% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Synchronize Folders}
\usage{
push(user = c("shane", "vlad"))
}
\arguments{
\item{user}{The path to the source folder.}
}
\value{
None
}
\description{
This function synchronizes files between a source folder and a destination folder.
If a file is modified or deleted in the source folder, it will be updated or deleted in the destination folder.
}
