% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushAuto_shane.R
\name{pushAuto_shane}
\alias{pushAuto_shane}
\title{Synchronize Two Folders}
\usage{
pushAuto_shane()
}
\description{
This function synchronizes two folders by mirroring the files and directories
from the source folder to the destination folder.
}
\examples{
synchronizeFolders("C:/Path/to/Source/Folder", "//RemoteComputer/Shared/Folder")
}
