% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushShane.R
\name{pushShane}
\alias{pushShane}
\title{Synchronize Folders}
\usage{
pushShane()
}
\value{
None
}
\description{
This function synchronizes files and directory structure between a source folder and a destination folder.
If a file is modified or deleted in the source folder, it will be updated or deleted in the destination folder.
The directory structure in the source folder will be replicated in the destination folder.
}
